requirements.txt
Flask-WTF==1.1.1
Flask-Login==0.6.3
email-validator==1.3.1
python-dotenv==1.0.1
itsdangerous==2.1.2
